Big update

+++ Todo.py
@@ -12,7 +12,14 @@ import os
 # Create a __main__
 # Factor and improve the code by turning it to functions and classes
 # Upgrade the todolist to a dictionnary so that i can contain more info about each task
-
+class notepad():
+    def __init__(self):
+        ''' the Core functionality of the script, can be modified to load or
+            create new notepads'''
+    
+    # TODO LATER
+    # FIRST learn how to append to an already existing todo
+    # IE load a notepad then try to implement the class.
 
 
 
@@ -23,17 +30,60 @@ import os
 # argslengh = len(sys.argv)
 # args = sys.argv
 
-def core():
+def main():
 
+    ExistingTodos = []
     ToDolist = []
-    numberoftodos = 0
     def Startscreen():
         print("Please select an argument. \n 1- add \n 2- remove \n 3- exit \n 4- show \n 5- args")
     number = 1
-
-
     print('Welcome to To Do list!')
-    Startscreen()   
+
+    # TODO: find a better implementation for this, folderName and subfolders are unused
+    for folderName, subfolders, filenames in os.walk("D:/Projects/TodoList Python"): # NOTE: May need to change these 2
+        for Todos in filenames:
+            numberrelatedtofiledetected = 0
+            if Todos.lower().endswith(".txt"):
+                ExistingTodos.append(Todos)
+                numberrelatedtofiledetected += 1
+                displayfiledetected = os.path.join('D:/Projects/TodoList Python', Todos) # Change this later                                                  
+                print(f'{numberrelatedtofiledetected}: {displayfiledetected}') # NOTE: May need to change these 2                         
+        
+        else:
+            break #
+
+    print('')
+    Loadorcreate = input('Would you like to load a todo list or create a new one? (l/n)')
+    while True:
+        if Loadorcreate == 'l':
+            for file in ExistingTodos:
+                print(file)
+
+            while True:
+                # TODO: add a gate for when someone inputs a 0
+                # TODO: add a gate for when there are no files and user picked load (l)
+                selectloadfileinput = int(input('which file would you like to open ?'))  - 1
+
+                if selectloadfileinput > len(ExistingTodos):
+                    print('That file does not exist! \n')
+
+                else:
+                    loadthisfile = os.path.join('D:/Projects/TodoList Python', str(ExistingTodos[selectloadfileinput]))
+                    with open(loadthisfile, 'r') as loadedfile:
+                        for line in loadedfile:
+                            ToDolist.append(line)
+                    break
+                break
+            break
+
+
+        elif Loadorcreate == 'n':
+            loadthisfile = False
+            break 
+
+
+    Startscreen()
+    # TODO: CREATE A CLASS FOR THIS FOR THE LOAD IMPLEMENTATION
     while True:    
         inputtedCommand = input('>> ')
 
@@ -97,14 +147,18 @@ def core():
                     Tosavefile = f'Save{number}.txt'
 
                 # Writes to a file
-                with open(Tosavefile, 'w') as f:
-                    for line in ToDolist:
-                        f.write(line + '\n')                        
+                if loadthisfile:
+                    with open(loadthisfile, 'w') as f:
+                            f.write('\n'.join([' '.join(i) for i in ToDolist]))
+                else: 
+                    with open(Tosavefile, 'w') as f:
+                        f.write('\n'.join([' '.join(i) for i in ToDolist]))
+
                 break
 
 
         elif inputtedCommand == 'show':
-            print('')
+            numberoftodos = 0
             # TODO: add a gate for when the list is empty
             for i in range(len(ToDolist)):
                 numberoftodos += 1
@@ -124,7 +178,7 @@ def core():
 
 
 if __name__ == '__main__':
-    core()
+    main()
 #print(" argument number: " + str(argnumber) + " arguments")
 #print('arguments detected: ' + argparsed)
